# docs/integration/openapi-iteration3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "教育服务系统 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/student/leave-requests": {
      "post": {
        "tags": [
          "学生服务"
        ],
        "summary": "提交请假申请 [P0]",
        "description": "学生提交请假申请。\n\n**业务规则：**\n- 双重逻辑外键校验：sys_user (user_type='student') + student_info (status='active')\n- end_time 必须晚于 start_time\n- 新建状态为 pending\n\n数据写入 student_admin_service 表 (service_type='leave')。",
        "operationId": "create_leave_request_api_v1_student_leave_requests_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "学生服务"
        ],
        "summary": "查询请假记录 [P1]",
        "operationId": "list_leaves_api_v1_student_leave_requests_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Student Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/leave-requests/{request_id}/approve": {
      "put": {
        "tags": [
          "学生服务"
        ],
        "summary": "审批请假 [P0]",
        "description": "班主任/管理员审批请假申请。\n\n**业务规则：**\n- 仅 pending 状态可审批\n- 条件 UPDATE WHERE status='pending' 防并发\n- 审批后状态变为 approved/rejected（终态，不可回退）",
        "operationId": "approve_leave_api_v1_student_leave_requests__request_id__approve_put",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveApprove"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/leave-requests/{request_id}/cancel": {
      "put": {
        "tags": [
          "学生服务"
        ],
        "summary": "撤销请假 [P1]",
        "description": "学生自主撤销待审批的请假申请。\n\n**业务规则：**\n- 仅 pending 状态可撤销\n- 仅本人可撤销\n- 撤销后状态变为 cancelled（终态）",
        "operationId": "cancel_leave_api_v1_student_leave_requests__request_id__cancel_put",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          },
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "学生ID",
              "title": "Student Id"
            },
            "description": "学生ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/feedback-tickets": {
      "post": {
        "tags": [
          "学生服务"
        ],
        "summary": "提交投诉/建议 [P0]",
        "description": "学生提交投诉、建议或咨询工单。\n\n**业务规则：**\n- 匿名投诉 student_id 可为 NULL\n- 非匿名时校验 student_id → sys_user (user_type='student')\n- 自动判定优先级（退款/律师 → urgent，投诉/紧急 → high）",
        "operationId": "create_feedback_api_v1_student_feedback_tickets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "学生服务"
        ],
        "summary": "查询投诉列表 [P1]",
        "operationId": "list_feedbacks_api_v1_student_feedback_tickets_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Student Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/feedback-tickets/{ticket_id}": {
      "put": {
        "tags": [
          "学生服务"
        ],
        "summary": "处理投诉 [P0]",
        "description": "处理投诉工单 — 更新状态、分配处理人、记录解决方案。\n\n**状态流转:** pending → processing → resolved → closed (终态)\nclosed 为终态，不可回退。",
        "operationId": "update_feedback_api_v1_student_feedback_tickets__ticket_id__put",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/psych/record": {
      "post": {
        "tags": [
          "学生服务"
        ],
        "summary": "记录心理交互 [P1] [Dify白名单]",
        "description": "保存 Dify AI 识别的学生情绪状态，联动更新心理画像，条件触发预警。\n\n**预警触发:** emotion_score < 20 或高危关键词 → high | 20-39 → medium | >= 40 → low\n**重要:** AI 只做风险识别，不做医学诊断。高危预警必须人工介入。",
        "operationId": "record_psych_api_v1_student_psych_record_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PsychRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/psych/alerts": {
      "get": {
        "tags": [
          "学生服务"
        ],
        "summary": "查询心理预警列表 [P1]",
        "operationId": "list_psych_alerts_api_v1_student_psych_alerts_get",
        "parameters": [
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Risk Level"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/psych/alerts/{alert_id}": {
      "put": {
        "tags": [
          "学生服务"
        ],
        "summary": "处理心理预警 [P1]",
        "description": "班主任/管理员处理心理预警。\n\n**状态流转:** pending → following → resolved (已解除) / dismissed (误报)",
        "operationId": "update_psych_alert_api_v1_student_psych_alerts__alert_id__put",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PsychAlertUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/applications": {
      "get": {
        "tags": [
          "学生服务"
        ],
        "summary": "查询留学申请进度 [P1] [Dify白名单]",
        "description": "查询学生的留学申请进度列表。数据来源: application_progress 表。",
        "operationId": "list_applications_api_v1_student_applications_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Student Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/student/scores": {
      "get": {
        "tags": [
          "学生服务"
        ],
        "summary": "查询学生成绩 [P1]",
        "description": "查询学生各课程成绩，支持按学期筛选。数据来源: student_score 表。",
        "operationId": "list_scores_api_v1_student_scores_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "学生ID",
              "title": "Student Id"
            },
            "description": "学生ID"
          },
          {
            "name": "semester",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "学期筛选，如 2025-2026-1",
              "title": "Semester"
            },
            "description": "学期筛选，如 2025-2026-1"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaveCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Student Id",
            "description": "学生ID → sys_user.id"
          },
          "service_type": {
            "type": "string",
            "title": "Service Type",
            "description": "服务类型: leave/exam_query/other",
            "default": "leave"
          },
          "leave_type": {
            "type": "string",
            "title": "Leave Type",
            "description": "请假类型: sick/personal/emergency"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time",
            "description": "开始时间 (YYYY-MM-DD HH:mm)"
          },
          "end_time": {
            "type": "string",
            "title": "End Time",
            "description": "结束时间 (YYYY-MM-DD HH:mm)"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "description": "申请事由"
          },
          "attachment_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment Url",
            "description": "附件URL"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "leave_type",
          "start_time",
          "end_time",
          "reason"
        ],
        "title": "LeaveCreate",
        "description": "提交请假申请 — POST /student/leave-requests"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "DifyServiceToken": {
        "type": "http",
        "description": "Dify HTTP 节点调用 FastAPI 白名单接口的服务令牌",
        "scheme": "bearer"
      }
    }
  }
}